% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{db_annotators}
\alias{db_annotators}
\title{Annotators in a PhysioBank database}
\usage{
db_annotators(db_name)
}
\arguments{
\item{db_name}{}
}
\value{
tibble
}
\description{
Annotators in a PhysioBank database
}
