% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{db_records}
\alias{db_records}
\title{Records in a PhysioBank database}
\usage{
db_records(db_name)
}
\arguments{
\item{db_name}{}
}
\value{
character vector
}
\description{
Records in a PhysioBank database
}
