% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ls_db}
\alias{ls_db}
\title{List entire file structure of a PhysioBank database}
\usage{
ls_db(db_name)
}
\arguments{
\item{db_name}{}
}
\value{
character vector
}
\description{
List entire file structure of a PhysioBank database
}
