% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_read_record.R
\name{wf_read_record}
\alias{wf_read_record}
\alias{wf_read_sample}
\alias{wf_read}
\alias{wf_read_header}
\title{Read a WFDB record and return the signal and record descriptors as
attributes in a Record or MultiRecord object.}
\usage{
wf_read_record(
  path,
  sample_from = 0L,
  sample_to = NULL,
  channels = NULL,
  physical = TRUE,
  m2s = TRUE,
  smooth_frames = TRUE,
  ignore_skew = FALSE,
  return_float_type = c(64L, 32L, 16L, 8L),
  force_channels = TRUE,
  channel_names = NULL,
  warn_empty = FALSE,
  ...
)

wf_read_sample(
  path,
  sample_from = 0L,
  sample_to = NULL,
  channels = NULL,
  return_float_type = c(64L, 32L, 16L, 8L),
  channel_names = NULL,
  warn_empty = FALSE,
  ...
)

wf_read(
  path,
  sample_from = 0L,
  sample_to = NULL,
  channels = NULL,
  physical = TRUE,
  m2s = TRUE,
  smooth_frames = TRUE,
  ignore_skew = FALSE,
  return_float_type = c(64L, 32L, 16L, 8L),
  force_channels = TRUE,
  channel_names = NULL,
  warn_empty = FALSE,
  ...
)

wf_read_header(path, rd_segments = TRUE, ...)
}
\arguments{
\item{path}{The name of the WFDB record to be read, without any
file extensions. If the argument contains any path delimiter characters,
the argument will be interpreted as PATH/BASE_RECORD.
Both relative and absolute paths are accepted.
If the pn_dir parameter is set, this parameter should contain
just the base record name, and the files fill be searched for
remotely. Otherwise, the data files will be searched for in the local path.}

\item{sample_from}{(integer or NULL)
The starting sample number to read for all channels; 0-indexed.}

\item{sample_to}{(integer or NULL)
The sample number at which to stop reading for
all channels. Reads the entire duration by default; 0-indexed}

\item{channels}{List of integer indices specifying the channels to be read.
Reads all channels by default.}

\item{physical}{Specifies whether to return signals in
physical units in the \code{p_signal} field (TRUE), or
digital units in the \code{d_signal} field (\code{FALSE}).}

\item{m2s}{Used when reading multi-segment records.
Specifies whether to directly return a WFDB \code{MultiRecord} object (\code{FALSE}),
or to convert it into and return a WFDB \code{Record} object (\code{TRUE}).}

\item{smooth_frames}{Specifies whether to smooth the samples in signals
with more than one sample per frame and return an (MxN) uniform numpy a
rray as the d_signal or p_signal field (True), or to return a list of
1d numpy arrays containing every expanded sample as the e_d_signal or
e_p_signal field (False).}

\item{ignore_skew}{Used when reading records with at least one skewed
signal. Specifies whether to apply the skew to align the signals in the
output variable (False), or to ignore the skew field and load in all
values contained in the dat files unaligned (True).}

\item{return_float_type}{The numpy array \code{dtype} of the returned signals.
Options are: 64, 32, 16, and 8, where the value represents the
numpy \code{int} or \code{float} \code{dtype.} Note that the value cannot be 8 when
physical is True since there is no float8 format.}

\item{force_channels}{Used when reading multi-segment variable layout
records. Whether to update the layout specification record,
and the converted Record object if m2s is True, to match the input
channels argument, or to omit channels in which no read segment
contains the signals.}

\item{channel_names}{List of channel names to return. If this
parameter is specified, it takes precedence over channels.}

\item{warn_empty}{Whether to display a warning if the specified channel
indices or names are not contained in the record, and no signal is returned.}

\item{rd_segments}{Used when reading multi-segment headers.
If \code{TRUE}, segment headers will also be read
(into the record object’s segments field).}
}
\value{
A \code{Record} output
}
\description{
Read a WFDB record and return the signal and record descriptors as
attributes in a Record or MultiRecord object.
}
\note{
https://wfdb.readthedocs.io/en/latest/io.html#module-wfdb.io
}
