% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb.R
\name{wfdb}
\alias{wfdb}
\title{The \code{wfdb} Python Module}
\usage{
wfdb()
}
\value{
A \code{python.builtin.module}
}
\description{
The \code{wfdb} Python Module
}
\examples{
\dontshow{if (have_wfdb()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
wfdb()
\dontshow{\}) # examplesIf}
}
