% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_wfdb.R
\name{install_wfdb}
\alias{install_wfdb}
\alias{have_wfdb}
\alias{wfdb_version}
\title{Install the \code{wfdb} Python Module}
\usage{
install_wfdb(..., packages = "wfdb")

have_wfdb()

wfdb_version()
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{packages}{additional packages to install.  If \code{wfdb} is missing,
it will be added}
}
\value{
Output of \link[reticulate:py_install]{reticulate::py_install}
}
\description{
Install the \code{wfdb} Python Module
}
\examples{
if (have_wfdb()) {
   wfdb_version()
}

}
